requirements.txt
Django~=5.0.7
djangorestframework~=3.15.2
python-dotenv~=1.0.1
typing_extensions~=4.12.2
pillow~=10.4.0
pip~=24.2
attrs~=23.2.0
pytz~=2024.1
PyYAML~=6.0.2
asgiref~=3.8.1
sqlparse~=0.5.0
psycopg2~=2.9.9
requests~=2.32.3
tablib~=3.5.0
packaging~=24.1
inflection~=0.5.1
uritemplate~=4.1.1
setuptools~=72.1.0
install~=1.3.5